+++ go/processtree/workernode.go
@@ -192,8 +192,8 @@ func (s *WorkerNode) doUnbootedState(monitor *WorkerMonitor) string { // -> {SBo
 		s.L.Lock()
 		parts := strings.Split(monitor.tree.ExecCommand, " ")
 		cmd := exec.Command(parts[0], parts[1:]...)
-		file := monitor.remoteMasterFile
-		cmd.Env = append(os.Environ(), fmt.Sprintf("ZEUS_MASTER_FD=%d", file.Fd()))
+		file := monitor.remoteCoordinatorFile
+		cmd.Env = append(os.Environ(), fmt.Sprintf("ZEUS_COORDINATOR_FD=%d", file.Fd()))
 		cmd.ExtraFiles = []*os.File{file}
 		go s.babysitRootProcess(cmd)
 		s.L.Unlock()